requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
bottle==0.12.25
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
Flask-Testing==0.8.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
monotonic==1.6
mysql-connector-python==8.3.0
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyparsing==3.1.4
pytest==8.3.3
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.6
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
toml==0.10.2
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
xxhash==3.4.1
